Refactor exception classes
* Moved `PostgresMessageMeta` and `PostgresMessage` into their own file
  to make it easier to add typings later on
* Re-exported `PostgresMessageMeta` and `PostgresMessage` from
  `asyncpg.exceptions._base` to maintain backwards compatibility
* Removed `import asyncpg` statements and import within methods to
  prevent circular imports
